Add ProductImage entity and update relationships
This commit introduces a new `ProductImage` entity and establishes its relationship with the `Product` entity in the `ApplicationDbContext`. It includes the addition of `DbSet<ProductImage>` and `DbSet<Product>` properties. The migration updates the database schema by renaming the table from `ProductImage` to `ProductImages` and sets up a foreign key relationship between `ProductImages` and `Products`. The `ApplicationDbContextModelSnapshot` is also updated to reflect these changes.